+++ CHANGELOG.md
@@ -37,6 +37,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Add optional chaining and explicitely return "Not found" page - #408 by @dominik-zeglen
 - Do not store errors in form component - #410 by @dominik-zeglen
 - Snapshot strict null mode error count - #411 by @dominik-zeglen
+- Check if number of errors in strict null check mode increased - #412 by @dominik-zeglen
 
 ## 2.0.0